requirements.txt
Django==5.0.1
django-unfold==0.38.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.6.0
Pillow==10.1.0
django-ckeditor==6.7.3
reportlab==4.0.9
